Easier script execution from user's private bin.
Use generic `.profile` login script to check if `bin` directory
exists in `$HOME`, and add to `$PATH` if so. Simplifies execution
of scripts in `$HOME/bin` by default, and fixes beargle/dotfiles#5.

`bash` reads `.profile` for backwards compatibility with `/bin/sh`,
so we use this instead of `.bashrc_profile` to support wider variety
of login shells.